# internal/docs/swagger.yaml
basePath: /v1
definitions:
...
  common.ErrorResponse:
    properties:
      code:
        type: string
      description:
        type: string
      msg:
        type: string
    type: object
...
  user.PolicyModel:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  user.UserResponse:
    properties:
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      facility:
        type: string
      id:
        type: string
      phone:
        type: string
      policies:
        items:
          $ref: '#/definitions/user.PolicyModel'
        type: array
      position:
        type: string
      team:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: BuddyのAPIサーバー
  title: Buddy-API
  version: "1.0"
paths:
...
  /facilities/{facility_id}/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: Facility ID
        in: path
        name: facility_id
        required: true
        type: string
      - description: Username
        in: query
        name: username
        type: string
      - description: Position
        in: query
        name: position
        type: string
      - description: Department
        in: query
        name: department
        type: string
      - description: Team
        in: query
        name: team
        type: string
      - description: Sort Field
        in: query
        name: sort_field
        type: string
      - description: Sort Order (asc or desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 施設IDに紐づくユーザーを取得する
      tags:
      - User
...
  /users/{user_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 単一のユーザーを取得する
      tags:
      - User
swagger: "2.0"